Add IconBlock component and use in FeaturesGeneral
Introduced a new Astro component, IconBlock, used for displaying icons paired with a heading and paragraph. This component is then utilized in the newly added FeaturesGeneral component, which encapsulates a feature section of the UI with a title, subtitle, image and a dynamic set of IconBlocks. This promotes reusability and visual consistency across the application.