Add French translation for various site components, data files, and navigation settings
This commit adds French translations to various components of the website. Specifically, a French Contact Us section has been created, as well as French versions of FAQs and feature data files. Additionally, French settings for the site navigation have been defined. This provides French-speaking users with a more convenient and accessible experience when using the site.